Implements core backend frontent communication
Uses ICS and getAppointments function
