+++ src/main/web/app/components/Calendar.js
@@ -1,55 +1,28 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import CalendarView from './CalendarView';
+import TimeView from './TimeView';
+import Day from './Day';
 
-export default class Calendar extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { dailyEvents: [[], [], [], [], [], []] };
-    const date = new Date();
-    console.log(props.getAppointments(
-      'https://rapla.dhbw-stuttgart.de/rapla?key=txB1FOi5xd1wUJBWuX8lJhGDUgtMSFmnKLgAG_NVMhA_bi91ugPaHvrpxD-lcejo&day=6&month=3&year=2017&next=%3E%3E',
-      date.getDay(), date.getMonth(), date.getYear(), this.onAjaxSuccess, this.onAjaxError,
-    ));
-  }
+const dayNames = [
+  'Montag',
+  'Dienstag',
+  'Mittwoch',
+  'Donnerstag',
+  'Freitag',
+  'Samstag',
+];
 
-  onAjaxSuccess = (response) => {
-    const data = JSON.parse(response);
-    this.setState({ dailyEvents: this.makeDays(this.parseDates(data)) });
-    // TODO What should happen when we receive new appointments?
-    console.log(data);
-  };
-
-  onAjaxError = (error) => {
-  // TODO What should happen on error?
-    console.error(error);
-  }
-
-  parseDates = (events) => {
-    events.forEach((el) => {
-      const curDate = el.date.split('.');
-      const startTime = el.startTime.split(':');
-      el.Date = new Date(curDate[2], curDate[1] - 1, curDate[0], startTime[0], startTime[1]);
-    });
-    return events;
-  }
-
-  makeDays = (events) => {
-    const dailyEvents = [[], [], [], [], [], []];
-    events.forEach((el) => {
-      const index = el.Date.getDay() - 1;
-      dailyEvents[index].push(el);
-    });
-    return dailyEvents;
-  }
-
-  render() {
-    return (
-      <CalendarView dailyEvents={this.state.dailyEvents} />
-    );
-  }
+export default function Calendar(props) {
+  return (
+    <container>
+      <div className="calendar">
+        <TimeView start={8} end={18} />
+        {dayNames.map((name, i) => <Day key={i} eventData={props.dailyEvents[i]} name={name} />)}
+      </div>
+    </container>
+  );
 }
 
 Calendar.propTypes = {
-  getAppointments: PropTypes.func.isRequired,
+  dailyEvents: PropTypes.arrayOf(PropTypes.arrayOf(PropTypes.object)).isRequired,
 };
