--- src/main/web/app/components/CalendarView.js
@@ -1,22 +1,0 @@
-import React from 'react';
-import TimeView from './TimeView';
-import Day from './Day';
-
-export default function CalendarView(props) {
-  const dayNames = [
-    'Montag',
-    'Dienstag',
-    'Mittwoch',
-    'Donnerstag',
-    'Freitag',
-    'Samstag',
-  ];
-  return (
-    <container>
-      <div className="calendar">
-        <TimeView start={8} end={18} />
-        {dayNames.map((name, i) => <Day key={i} eventData={props.dailyEvents[i]} name={name} />)}
-      </div>
-    </container>
-  );
-}
