+++ src/main/web/app/scss/calendar.scss
@@ -24,8 +24,6 @@ container {
   justify-content: flex-start;
 }
 
-
-
 .day {
   position: relative;
   box-sizing: border-box;
@@ -111,7 +109,7 @@ container {
     }
 
     &--header {
-      font-size: 1.4em;
+      font-size: 1.1em;
       margin: 15px 0;
     }
 
@@ -184,7 +182,7 @@ container {
     }
 
     &--header {
-      font-size: 1.2em;
+      font-size: 1.1em;
       margin: 15px 0 5px 0;
     }
